Move Data into visit
